minor refactoring

+++ requests/delete.php
@@ -4,7 +4,11 @@ switch (URL) {
         $api->deleteUser();
         break;
     case "posts":
-        $post_id = $_GET['id'];
-        $api->deletePost($post_id);
+        if(isset($_GET['id'])) {
+            $post_id = $_GET['id'];
+            $api->deletePost($post_id);
+        } else {
+            $api->throwError(405, "Post not found.");
+        }
         break;
 }
